tools/AutoLocalization/requirements.txt
setuptools>=67.8.0
wheel>=0.40.0
Cython>=0.29.25
lxml~=4.9.2
openai~=0.27.7
python-dotenv~=1.0.0
cchardet~=2.1.7
xmldiff>=2.6.3
opencc~=1.1.1
